% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kclass.R
\name{Test_Statistic}
\alias{Test_Statistic}
\title{Test-statistic evaluation}
\usage{
Test_Statistic(alpha, A, Z, Y)
}
\arguments{
\item{alpha}{coefficient/parameter to evaluate the test-statistic in (vector)}

\item{A}{exogenous variables (vector/matrix/data.frame)}

\item{Z}{included exogenous and endogenous regressors (vector/matrix/data.frame)}

\item{Y}{target variable (vector/matrix/data.frame)}
}
\value{
the test-statistic evaluated in the coefficient input alpha
}
\description{
Test-statistic evaluation
}
